optional-requirements.txt
# Format: #/name defines a new extras_require group called name
# Uncommented lines after the group definition insert things into that group.

#/e2be
python-olm>=3,<4
pycryptodome>=3,<4
unpaddedbase64>=1,<2

#/login
pillow>=4,<8
qrcode>=6,<7
